baseStation/requirements.txt
coverage==4.5.2
nose==1.3.7
numpy==1.16.1
opencv-contrib-python==4.0.0.21
opencv-python==4.0.0.21
Pillow==5.4.1
pyserial==3.4
